requirements.txt
# Medical Image Anonymization Pipeline - Python Dependencies

# Core Scientific Computing
numpy>=1.17,<2.0  # CRITICAL: EasyOCR requires numpy<2.0

# Image Processing
opencv-python>=4.5.0
Pillow>=9.0.0

# Medical Imaging
pydicom>=2.3.0
python-gdcm>=3.0.0  # DICOM decompression (no numpy conflict)

# OCR Engines
paddleocr>=2.6.0
easyocr>=1.6.0

# Deep Learning & AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=5.2.0
gdown>=5.0.0
torchxrayvision>=1.3.0  # multi-dataset chest X-ray pathology models
h5py>=3.0.0             # legacy: only used for the optional Keras→PyTorch converter

# Utilities
scikit-learn>=1.0.0

# Storage & Configuration
minio>=7.1.0
python-dotenv>=1.0.0

# FastAPI Backend
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# NOTE: pylibjpeg and pylibjpeg-libjpeg removed due to numpy>=2.0 conflict
# Use python-gdcm instead for compressed DICOM support (JPEG Lossless, JPEG 2000, etc.)
